search3/requirements.txt
django==4.2.16
djangorestframework==3.15.2
pandas==2.0.3
numpy==1.24.3
sentence-transformers==3.2.0
neo4j==5.24.0
gunicorn==23.0.0